Made tests parameterized
