+++ src/test/resources/SimpleMathProblems.csv
@@ -1,0 +1,15 @@
+input,expected
+1+1,2
+1e10,10000000000
+1E10,10000000000
+123.456E3,123456
+2.5e0,2.5
+1e-10,0.0000000001
+1E-10,0.0000000001,
+2135E-4,0.2135
+sqrt(152.399025e6),12345
+sin(3.e1),0.5
+sin( 3.e1),0.5
+sin(3.e1 ),0.5
+sin( 3.e1 ),0.5
+2.2e-16 * 10.2,2.244E-15
